refactor(utilities): separating the frontend on logic with UI.
